chore: Move setTeamIngestedEvent back
